refactor: create curve-features package
for now the onboard-helpers are transformed into a connect-wallet feature
