+++ apps/lend/tsconfig.json
@@ -8,7 +8,7 @@
       "@/ui/*": ["../../../packages/ui/src/*"],
       "@/images": ["../../../packages/ui/src/images/index.ts"],
       "@/images/*": ["../../../packages/ui/src/images/*"],
-      "@/onboard": ["../../../packages/onboard-helpers/src/index.ts"],
+      "@/shared/features/*": ["../../../packages/curve-features/src/*"],
       "@/shared/curve-lib": ["../../../packages/curve-lib/src/index.ts"],
       "@/shared/api/*": ["../../../packages/curve-lib/src/shared/api/*"],
       "@/shared/lib/*": ["../../../packages/curve-lib/src/shared/lib/*"],
