+++ apps/loan/package.json
@@ -30,6 +30,7 @@
     "@supercharge/promise-pool": "^2.3.2",
     "bignumber.js": "^9.0.1",
     "cross-fetch": "^3.1.5",
+    "curve-features": "*",
     "curve-lib": "*",
     "dayjs": "^1.11.7",
     "ethers": "^5",
@@ -43,7 +44,6 @@
     "next": "^13.5.6",
     "next-images": "^1.8.5",
     "numbro": "^2.4.0",
-    "onboard-helpers": "*",
     "react": "*",
     "react-aria": "3.22.0",
     "react-dom": "*",
