+++ apps/loan/src/layout/Header.tsx
@@ -10,7 +10,7 @@ import { DEFAULT_LOCALES } from '@/lib/i18n'
 import { getLocaleFromUrl, getNetworkFromUrl, getPath, getRestFullPathname } from '@/utils/utilsRouter'
 import { getWalletSignerAddress } from '@/store/createWalletSlice'
 import { _parseRouteAndIsActive, formatNumber, isLoading } from '@/ui/utils'
-import { useConnectWallet } from '@/onboard'
+import { useConnectWallet } from '@/shared/features/connect-wallet'
 import { visibleNetworksList } from '@/networks'
 import useLayoutHeight from '@/hooks/useLayoutHeight'
 import useStore from '@/store/useStore'
