+++ apps/main/next.config.js
@@ -27,7 +27,7 @@ const nextConfiguration = {
     unoptimized: true,
   },
   trailingSlash: true,
-  transpilePackages: ['onboard-helpers', 'ui'],
+  transpilePackages: ['curve-features', 'curve-ui-kit'],
   webpack(config) {
     config.module.rules.push({
       test: /\.svg$/,
