+++ packages/curve-features/global.d.ts
