+++ packages/curve-features/package.json
