+++ packages/curve-features/src/connect-wallet/index.ts
@@ -1,0 +1,2 @@
+export * from './lib'
+export { initOnboard } from './lib/init'
