+++ packages/curve-features/src/connect-wallet/lib/index.ts
@@ -1,0 +1,7 @@
+export { connectWalletLocales } from './locales'
+export {
+  useConnectWallet,
+  useSetChain,
+  useSetLocale
+} from '@web3-onboard/react'
+export { initOnboard } from './init'
