+++ packages/curve-features/src/connect-wallet/lib/init.ts
@@ -1,21 +1,19 @@
-import * as onboard from '@web3-onboard/react'
 import {
-  injected,
-  trezor,
-  ledger,
-  gnosis,
+  bitget,
   coinbaseWalletSdk,
   fortmatic,
-  walletConnect,
-  torus,
-  phantom,
   frontier,
-  bitget,
-  trust,
+  gnosis,
+  injected,
+  ledger,
   metamaskSDKWallet,
+  phantom,
+  torus,
+  trezor,
+  trust,
+  walletConnect
 } from './utils/walletModules'
-
-export * from '@web3-onboard/react'
+import * as onboard from '@web3-onboard/react'
 
 export function initOnboard(i18n: any, locale: any, themeType: any, networks: any) {
   let theme: 'system' | 'light' | 'dark' = 'system'
@@ -31,7 +29,7 @@ export function initOnboard(i18n: any, locale: any, themeType: any, networks: an
       id: network.hex,
       token: network.symbol,
       label: network.name,
-      rpcUrl: network.rpcUrlConnectWallet,
+      rpcUrl: network.rpcUrlConnectWallet
     }
   })
 
@@ -49,35 +47,36 @@ export function initOnboard(i18n: any, locale: any, themeType: any, networks: an
       frontier,
       bitget,
       trust,
-      metamaskSDKWallet,
+      metamaskSDKWallet
     ],
     chains,
     appMetadata: {
       name: 'Curve',
       description: 'Efficient stablecoin and non-stablecoin swapping',
-      icon: 'https://classic.curve.fi/logo-square.svg',
+      icon: 'https://classic.curve.fi/logo-square.svg'
     },
     disableFontDownload: true,
     i18n,
     notify: {
       desktop: {
         enabled: true,
-        transactionHandler: () => {},
-        position: 'topRight',
-      },
+        transactionHandler: () => {
+        },
+        position: 'topRight'
+      }
     },
     theme,
     accountCenter: {
       desktop: {
-        enabled: false,
+        enabled: false
       },
       mobile: {
-        enabled: false,
-      },
+        enabled: false
+      }
     },
     connect: {
-      autoConnectLastWallet: false,
-    },
+      autoConnectLastWallet: false
+    }
   })
 
   if (walletState && locale) {
