+++ packages/curve-features/src/connect-wallet/lib/locales/index.ts
@@ -1,0 +1,4 @@
+import zhHans from './zh-Hans'
+import zhHant from './zh-Hant'
+
+export const connectWalletLocales = { 'zh-Hans': zhHans, 'zh-Hant': zhHant }
