+++ packages/curve-features/src/connect-wallet/lib/locales/zh-Hant.ts
