+++ packages/curve-features/src/connect-wallet/lib/utils/walletModules.ts
